+++ Objects/MapObject.cs
@@ -3,8 +3,8 @@ namespace Talos.Objects
 {
     internal abstract class MapObject
     {
-        internal virtual byte SourceX { get; set; }
-        internal virtual byte SourceY { get; set; }
+        internal virtual short SourceX { get; set; }
+        internal virtual short SourceY { get; set; }
         internal virtual short SourceMapID { get; set; }
     }
 }
